refactor: persist todos and keep page as server component
Address AO review findings:

- Persist todos to localStorage so a refresh or tab close no longer wipes
  the list. Load via a "storage" event subscription (also syncs across
  tabs) and persist on every mutation through a setTodos wrapper.
- Move the interactive widget into a new client component (app/todo-app.tsx)
  so app/page.tsx can stay a server component. The route's static markup
  (main/heading) is no longer shipped in the client bundle.
